Crate icu_datetime_data

Data for the icu_datetime crate

This data was generated with CLDR version 48.1.0, ICU version release-78.1rc, and LSTM segmenter version v0.1.0.

Macros

impl_datetime_names_dayperiod_v1

Implement DataProvider<DatetimeNamesDayperiodV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_buddhist_v1

Implement DataProvider<DatetimeNamesMonthBuddhistV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_chinese_v1

Implement DataProvider<DatetimeNamesMonthChineseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_coptic_v1

Implement DataProvider<DatetimeNamesMonthCopticV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_dangi_v1

Implement DataProvider<DatetimeNamesMonthDangiV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_ethiopian_v1

Implement DataProvider<DatetimeNamesMonthEthiopianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_gregorian_v1

Implement DataProvider<DatetimeNamesMonthGregorianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_hebrew_v1

Implement DataProvider<DatetimeNamesMonthHebrewV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_hijri_v1

Implement DataProvider<DatetimeNamesMonthHijriV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_indian_v1

Implement DataProvider<DatetimeNamesMonthIndianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_japanese_v1

Implement DataProvider<DatetimeNamesMonthJapaneseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_japanext_v1

Implement DataProvider<DatetimeNamesMonthJapanextV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_persian_v1

Implement DataProvider<DatetimeNamesMonthPersianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_month_roc_v1

Implement DataProvider<DatetimeNamesMonthRocV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_weekday_v1

Implement DataProvider<DatetimeNamesWeekdayV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_buddhist_v1

Implement DataProvider<DatetimeNamesYearBuddhistV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_chinese_v1

Implement DataProvider<DatetimeNamesYearChineseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_coptic_v1

Implement DataProvider<DatetimeNamesYearCopticV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_dangi_v1

Implement DataProvider<DatetimeNamesYearDangiV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_ethiopian_v1

Implement DataProvider<DatetimeNamesYearEthiopianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_gregorian_v1

Implement DataProvider<DatetimeNamesYearGregorianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_hebrew_v1

Implement DataProvider<DatetimeNamesYearHebrewV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_hijri_v1

Implement DataProvider<DatetimeNamesYearHijriV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_indian_v1

Implement DataProvider<DatetimeNamesYearIndianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_japanese_v1

Implement DataProvider<DatetimeNamesYearJapaneseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_japanext_v1

Implement DataProvider<DatetimeNamesYearJapanextV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_persian_v1

Implement DataProvider<DatetimeNamesYearPersianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_names_year_roc_v1

Implement DataProvider<DatetimeNamesYearRocV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_buddhist_v1

Implement DataProvider<DatetimePatternsDateBuddhistV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_chinese_v1

Implement DataProvider<DatetimePatternsDateChineseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_coptic_v1

Implement DataProvider<DatetimePatternsDateCopticV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_dangi_v1

Implement DataProvider<DatetimePatternsDateDangiV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_ethiopian_v1

Implement DataProvider<DatetimePatternsDateEthiopianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_gregorian_v1

Implement DataProvider<DatetimePatternsDateGregorianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_hebrew_v1

Implement DataProvider<DatetimePatternsDateHebrewV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_hijri_v1

Implement DataProvider<DatetimePatternsDateHijriV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_indian_v1

Implement DataProvider<DatetimePatternsDateIndianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_japanese_v1

Implement DataProvider<DatetimePatternsDateJapaneseV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_japanext_v1

Implement DataProvider<DatetimePatternsDateJapanextV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_persian_v1

Implement DataProvider<DatetimePatternsDatePersianV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_date_roc_v1

Implement DataProvider<DatetimePatternsDateRocV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_glue_v1

Implement DataProvider<DatetimePatternsGlueV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_time_v1

Implement DataProvider<DatetimePatternsTimeV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_cities_override_v1

Implement DataProvider<TimezoneNamesCitiesOverrideV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_cities_root_v1

Implement DataProvider<TimezoneNamesCitiesRootV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_essentials_v1

Implement DataProvider<TimezoneNamesEssentialsV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_generic_long_v1

Implement DataProvider<TimezoneNamesGenericLongV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_generic_short_v1

Implement DataProvider<TimezoneNamesGenericShortV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_locations_override_v1

Implement DataProvider<TimezoneNamesLocationsOverrideV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_locations_root_v1

Implement DataProvider<TimezoneNamesLocationsRootV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_specific_long_v1

Implement DataProvider<TimezoneNamesSpecificLongV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_specific_short_v1

Implement DataProvider<TimezoneNamesSpecificShortV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_timezone_names_standard_long_v1

Implement DataProvider<TimezoneNamesStandardLongV1> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.