Crate icu_datetime_data

Data for the icu_datetime crate

This data was generated with CLDR version 45.0.0, ICU version icu4x/2024-05-16/75.x, and LSTM segmenter version v0.1.0.

Macros

impl_datetime_buddhist_datelengths_v1

Implement DataProvider<BuddhistDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_buddhist_datesymbols_v1

Implement DataProvider<BuddhistDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_chinese_datelengths_v1

Implement DataProvider<ChineseDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_chinese_datesymbols_v1

Implement DataProvider<ChineseDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_coptic_datelengths_v1

Implement DataProvider<CopticDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_coptic_datesymbols_v1

Implement DataProvider<CopticDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_dangi_datelengths_v1

Implement DataProvider<DangiDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_dangi_datesymbols_v1

Implement DataProvider<DangiDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_ethiopic_datelengths_v1

Implement DataProvider<EthiopianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_ethiopic_datesymbols_v1

Implement DataProvider<EthiopianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_gregory_datelengths_v1

Implement DataProvider<GregorianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_gregory_datesymbols_v1

Implement DataProvider<GregorianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_hebrew_datelengths_v1

Implement DataProvider<HebrewDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_hebrew_datesymbols_v1

Implement DataProvider<HebrewDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_indian_datelengths_v1

Implement DataProvider<IndianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_indian_datesymbols_v1

Implement DataProvider<IndianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_islamic_datelengths_v1

Implement DataProvider<IslamicDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_islamic_datesymbols_v1

Implement DataProvider<IslamicDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_japanese_datelengths_v1

Implement DataProvider<JapaneseDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_japanese_datesymbols_v1

Implement DataProvider<JapaneseDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_japanext_datelengths_v1

Implement DataProvider<JapaneseExtendedDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_japanext_datesymbols_v1

Implement DataProvider<JapaneseExtendedDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_buddhist_date_v1

Implement DataProvider<BuddhistDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_buddhist_skeleton_v1

Implement DataProvider<BuddhistDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_chinese_date_v1

Implement DataProvider<ChineseDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_chinese_skeleton_v1

Implement DataProvider<ChineseDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_coptic_date_v1

Implement DataProvider<CopticDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_coptic_skeleton_v1

Implement DataProvider<CopticDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_dangi_date_v1

Implement DataProvider<DangiDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_dangi_skeleton_v1

Implement DataProvider<DangiDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_datetime_v1

Implement DataProvider<DateTimePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_ethiopic_date_v1

Implement DataProvider<EthiopianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_ethiopic_skeleton_v1

Implement DataProvider<EthiopianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_gregory_date_v1

Implement DataProvider<GregorianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_gregory_skeleton_v1

Implement DataProvider<GregorianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_hebrew_date_v1

Implement DataProvider<HebrewDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_hebrew_skeleton_v1

Implement DataProvider<HebrewDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_indian_date_v1

Implement DataProvider<IndianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_indian_skeleton_v1

Implement DataProvider<IndianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_islamic_date_v1

Implement DataProvider<IslamicDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_islamic_skeleton_v1

Implement DataProvider<IslamicDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_japanese_date_v1

Implement DataProvider<JapaneseDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_japanese_skeleton_v1

Implement DataProvider<JapaneseDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_japanext_date_v1

Implement DataProvider<JapaneseExtendedDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_japanext_skeleton_v1

Implement DataProvider<JapaneseExtendedDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_persian_date_v1

Implement DataProvider<PersianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_persian_skeleton_v1

Implement DataProvider<PersianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_roc_date_v1

Implement DataProvider<RocDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_roc_skeleton_v1

Implement DataProvider<RocDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_time_skeleton_v1

Implement DataProvider<TimeNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_patterns_time_v1

Implement DataProvider<TimePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_persian_datelengths_v1

Implement DataProvider<PersianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_persian_datesymbols_v1

Implement DataProvider<PersianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_roc_datelengths_v1

Implement DataProvider<RocDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_roc_datesymbols_v1

Implement DataProvider<RocDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_skeletons_v1

Implement DataProvider<DateSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_buddhist_months_v1

Implement DataProvider<BuddhistMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_buddhist_years_v1

Implement DataProvider<BuddhistYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_chinese_months_v1

Implement DataProvider<ChineseMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_chinese_years_v1

Implement DataProvider<ChineseYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_coptic_months_v1

Implement DataProvider<CopticMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_coptic_years_v1

Implement DataProvider<CopticYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_dangi_months_v1

Implement DataProvider<DangiMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_dangi_years_v1

Implement DataProvider<DangiYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_dayperiods_v1

Implement DataProvider<DayPeriodNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_ethiopic_months_v1

Implement DataProvider<EthiopianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_ethiopic_years_v1

Implement DataProvider<EthiopianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_gregory_months_v1

Implement DataProvider<GregorianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_gregory_years_v1

Implement DataProvider<GregorianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_hebrew_months_v1

Implement DataProvider<HebrewMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_hebrew_years_v1

Implement DataProvider<HebrewYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_indian_months_v1

Implement DataProvider<IndianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_indian_years_v1

Implement DataProvider<IndianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_islamic_months_v1

Implement DataProvider<IslamicMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_islamic_years_v1

Implement DataProvider<IslamicYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_japanese_months_v1

Implement DataProvider<JapaneseMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_japanese_years_v1

Implement DataProvider<JapaneseYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_japanext_months_v1

Implement DataProvider<JapaneseExtendedMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_japanext_years_v1

Implement DataProvider<JapaneseExtendedYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_persian_months_v1

Implement DataProvider<PersianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_persian_years_v1

Implement DataProvider<PersianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_roc_months_v1

Implement DataProvider<RocMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_roc_years_v1

Implement DataProvider<RocYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_symbols_weekdays_v1

Implement DataProvider<WeekdayNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_timelengths_v1

Implement DataProvider<TimeLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_timesymbols_v1

Implement DataProvider<TimeSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_exemplar_cities_v1

Implement DataProvider<ExemplarCitiesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_formats_v1

Implement DataProvider<TimeZoneFormatsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_generic_long_v1

Implement DataProvider<MetazoneGenericNamesLongV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_generic_short_v1

Implement DataProvider<MetazoneGenericNamesShortV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_specific_long_v1

Implement DataProvider<MetazoneSpecificNamesLongV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_time_zone_specific_short_v1

Implement DataProvider<MetazoneSpecificNamesShortV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impliterable_datetime_buddhist_datelengths_v1

Implement IterableDataProvider<BuddhistDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_buddhist_datesymbols_v1

Implement IterableDataProvider<BuddhistDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_chinese_datelengths_v1

Implement IterableDataProvider<ChineseDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_chinese_datesymbols_v1

Implement IterableDataProvider<ChineseDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_coptic_datelengths_v1

Implement IterableDataProvider<CopticDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_coptic_datesymbols_v1

Implement IterableDataProvider<CopticDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_dangi_datelengths_v1

Implement IterableDataProvider<DangiDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_dangi_datesymbols_v1

Implement IterableDataProvider<DangiDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_ethiopic_datelengths_v1

Implement IterableDataProvider<EthiopianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_ethiopic_datesymbols_v1

Implement IterableDataProvider<EthiopianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_gregory_datelengths_v1

Implement IterableDataProvider<GregorianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_gregory_datesymbols_v1

Implement IterableDataProvider<GregorianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_hebrew_datelengths_v1

Implement IterableDataProvider<HebrewDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_hebrew_datesymbols_v1

Implement IterableDataProvider<HebrewDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_indian_datelengths_v1

Implement IterableDataProvider<IndianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_indian_datesymbols_v1

Implement IterableDataProvider<IndianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_islamic_datelengths_v1

Implement IterableDataProvider<IslamicDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_islamic_datesymbols_v1

Implement IterableDataProvider<IslamicDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_japanese_datelengths_v1

Implement IterableDataProvider<JapaneseDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_japanese_datesymbols_v1

Implement IterableDataProvider<JapaneseDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_japanext_datelengths_v1

Implement IterableDataProvider<JapaneseExtendedDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_japanext_datesymbols_v1

Implement IterableDataProvider<JapaneseExtendedDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_buddhist_date_v1

Implement IterableDataProvider<BuddhistDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_buddhist_skeleton_v1

Implement IterableDataProvider<BuddhistDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_chinese_date_v1

Implement IterableDataProvider<ChineseDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_chinese_skeleton_v1

Implement IterableDataProvider<ChineseDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_coptic_date_v1

Implement IterableDataProvider<CopticDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_coptic_skeleton_v1

Implement IterableDataProvider<CopticDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_dangi_date_v1

Implement IterableDataProvider<DangiDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_dangi_skeleton_v1

Implement IterableDataProvider<DangiDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_datetime_v1

Implement IterableDataProvider<DateTimePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_ethiopic_date_v1

Implement IterableDataProvider<EthiopianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_ethiopic_skeleton_v1

Implement IterableDataProvider<EthiopianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_gregory_date_v1

Implement IterableDataProvider<GregorianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_gregory_skeleton_v1

Implement IterableDataProvider<GregorianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_hebrew_date_v1

Implement IterableDataProvider<HebrewDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_hebrew_skeleton_v1

Implement IterableDataProvider<HebrewDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_indian_date_v1

Implement IterableDataProvider<IndianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_indian_skeleton_v1

Implement IterableDataProvider<IndianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_islamic_date_v1

Implement IterableDataProvider<IslamicDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_islamic_skeleton_v1

Implement IterableDataProvider<IslamicDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_japanese_date_v1

Implement IterableDataProvider<JapaneseDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_japanese_skeleton_v1

Implement IterableDataProvider<JapaneseDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_japanext_date_v1

Implement IterableDataProvider<JapaneseExtendedDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_japanext_skeleton_v1

Implement IterableDataProvider<JapaneseExtendedDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_persian_date_v1

Implement IterableDataProvider<PersianDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_persian_skeleton_v1

Implement IterableDataProvider<PersianDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_roc_date_v1

Implement IterableDataProvider<RocDatePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_roc_skeleton_v1

Implement IterableDataProvider<RocDateNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_time_skeleton_v1

Implement IterableDataProvider<TimeNeoSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_patterns_time_v1

Implement IterableDataProvider<TimePatternV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_persian_datelengths_v1

Implement IterableDataProvider<PersianDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_persian_datesymbols_v1

Implement IterableDataProvider<PersianDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_roc_datelengths_v1

Implement IterableDataProvider<RocDateLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_roc_datesymbols_v1

Implement IterableDataProvider<RocDateSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_skeletons_v1

Implement IterableDataProvider<DateSkeletonPatternsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_buddhist_months_v1

Implement IterableDataProvider<BuddhistMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_buddhist_years_v1

Implement IterableDataProvider<BuddhistYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_chinese_months_v1

Implement IterableDataProvider<ChineseMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_chinese_years_v1

Implement IterableDataProvider<ChineseYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_coptic_months_v1

Implement IterableDataProvider<CopticMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_coptic_years_v1

Implement IterableDataProvider<CopticYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_dangi_months_v1

Implement IterableDataProvider<DangiMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_dangi_years_v1

Implement IterableDataProvider<DangiYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_dayperiods_v1

Implement IterableDataProvider<DayPeriodNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_ethiopic_months_v1

Implement IterableDataProvider<EthiopianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_ethiopic_years_v1

Implement IterableDataProvider<EthiopianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_gregory_months_v1

Implement IterableDataProvider<GregorianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_gregory_years_v1

Implement IterableDataProvider<GregorianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_hebrew_months_v1

Implement IterableDataProvider<HebrewMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_hebrew_years_v1

Implement IterableDataProvider<HebrewYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_indian_months_v1

Implement IterableDataProvider<IndianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_indian_years_v1

Implement IterableDataProvider<IndianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_islamic_months_v1

Implement IterableDataProvider<IslamicMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_islamic_years_v1

Implement IterableDataProvider<IslamicYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_japanese_months_v1

Implement IterableDataProvider<JapaneseMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_japanese_years_v1

Implement IterableDataProvider<JapaneseYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_japanext_months_v1

Implement IterableDataProvider<JapaneseExtendedMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_japanext_years_v1

Implement IterableDataProvider<JapaneseExtendedYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_persian_months_v1

Implement IterableDataProvider<PersianMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_persian_years_v1

Implement IterableDataProvider<PersianYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_roc_months_v1

Implement IterableDataProvider<RocMonthNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_roc_years_v1

Implement IterableDataProvider<RocYearNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_symbols_weekdays_v1

Implement IterableDataProvider<WeekdayNamesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_timelengths_v1

Implement IterableDataProvider<TimeLengthsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_timesymbols_v1

Implement IterableDataProvider<TimeSymbolsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_exemplar_cities_v1

Implement IterableDataProvider<ExemplarCitiesV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_formats_v1

Implement IterableDataProvider<TimeZoneFormatsV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_generic_long_v1

Implement IterableDataProvider<MetazoneGenericNamesLongV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_generic_short_v1

Implement IterableDataProvider<MetazoneGenericNamesShortV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_specific_long_v1

Implement IterableDataProvider<MetazoneSpecificNamesLongV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_time_zone_specific_short_v1

Implement IterableDataProvider<MetazoneSpecificNamesShortV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.